memory arithmetics classwork
