+++ classwork/CMakeLists.txt
@@ -2,5 +2,6 @@
 
 add_executable (classwork "main.cpp" "main.h")
 add_executable (vehicles "vehicle.cpp" "vehicle.h")
+add_executable (memarith "memarith.cpp" "memarith.h")
 
 # TODO: Add tests and install targets if needed.
